requirements.txt
mcp>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-ads>=25.0.0
google-shopping-merchant-accounts>=0.1.0
google-shopping-merchant-products>=0.1.0
google-shopping-merchant-inventories>=0.1.0
